+++ src/ControlSystem/Measurements/BothHorizons.hpp
@@ -10,8 +10,8 @@
 #include "ApparentHorizons/Callbacks/FindApparentHorizon.hpp"
 #include "ApparentHorizons/ComputeHorizonVolumeQuantities.hpp"
 #include "ApparentHorizons/HorizonAliases.hpp"
+#include "ApparentHorizons/InterpolationTarget.hpp"
 #include "ApparentHorizons/ObserveCenters.hpp"
-#include "ApparentHorizons/Target.hpp"
 #include "ControlSystem/Protocols/Measurement.hpp"
 #include "ControlSystem/Protocols/Submeasurement.hpp"
 #include "ControlSystem/RunCallbacks.hpp"
