+++ tests/Unit/ApparentHorizons/Test_InterpolationTargetApparentHorizon.cpp
@@ -10,7 +10,7 @@
 #include <vector>
 
 #include "ApparentHorizons/FastFlow.hpp"
-#include "ApparentHorizons/Target.hpp"
+#include "ApparentHorizons/InterpolationTarget.hpp"
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "Domain/BlockLogicalCoordinates.hpp"
